refactor: rename functions and files
