+++ pyproject.toml
@@ -39,7 +39,6 @@ select = [
     "E4",
     "E7",
     "E9",
-    "E999",
     "F",
     "UP",
     "I",
