+++ tests/test_lexer.py
@@ -34,7 +34,7 @@ class TestLexerChapter01(unittest.TestCase):
             gcc_preprocess(path, preprocessed_file.name)
             tokens = [
                 (token.type, token.value)
-                for token in lexer.run_lexer(preprocessed_file.name)
+                for token in lexer.tokenize(preprocessed_file.name)
             ]
 
         self.assertListEqual(tokens, expected_list)
@@ -58,7 +58,7 @@ class TestLexerChapter01(unittest.TestCase):
             gcc_preprocess(path, preprocessed_file.name)
             tokens = [
                 (token.type, token.value)
-                for token in lexer.run_lexer(preprocessed_file.name)
+                for token in lexer.tokenize(preprocessed_file.name)
             ]
 
         self.assertListEqual(tokens, expected_list)
@@ -84,7 +84,7 @@ class TestLexerChapter01(unittest.TestCase):
             except subprocess.CalledProcessError as e:
                 return e.returncode
 
-            tokens = lexer.run_lexer(preprocessed_file.name)
+            tokens = lexer.tokenize(preprocessed_file.name)
 
         for i in range(len(tokens)):
             tokens[i] = (tokens[i].type, tokens[i].value)
@@ -110,7 +110,7 @@ class TestLexerChapter01(unittest.TestCase):
             gcc_preprocess(path, preprocessed_file.name)
             tokens = [
                 (token.type, token.value)
-                for token in lexer.run_lexer(preprocessed_file.name)
+                for token in lexer.tokenize(preprocessed_file.name)
             ]
 
         self.assertListEqual(tokens, expected_list)
@@ -134,7 +134,7 @@ class TestLexerChapter01(unittest.TestCase):
             gcc_preprocess(path, preprocessed_file.name)
             tokens = [
                 (token.type, token.value)
-                for token in lexer.run_lexer(preprocessed_file.name)
+                for token in lexer.tokenize(preprocessed_file.name)
             ]
 
         self.assertListEqual(tokens, expected_list)
@@ -158,7 +158,7 @@ class TestLexerChapter01(unittest.TestCase):
             gcc_preprocess(path, preprocessed_file.name)
             tokens = [
                 (token.type, token.value)
-                for token in lexer.run_lexer(preprocessed_file.name)
+                for token in lexer.tokenize(preprocessed_file.name)
             ]
 
         self.assertListEqual(tokens, expected_list)
@@ -169,7 +169,7 @@ class TestLexerChapter01(unittest.TestCase):
         with NamedTemporaryFile(suffix=".i") as preprocessed_file:
             gcc_preprocess(path, preprocessed_file.name)
             with self.assertRaises(ValueError):
-                lexer.run_lexer(preprocessed_file.name)
+                lexer.tokenize(preprocessed_file.name)
 
     def test_backslash(self):
         path = os.path.join(self.test_samples_path, "invalid_lex", "backslash.c")
@@ -177,7 +177,7 @@ class TestLexerChapter01(unittest.TestCase):
         with NamedTemporaryFile(suffix=".i") as preprocessed_file:
             gcc_preprocess(path, preprocessed_file.name)
             with self.assertRaises(ValueError):
-                lexer.run_lexer(preprocessed_file.name)
+                lexer.tokenize(preprocessed_file.name)
 
     def test_backtick(self):
         path = os.path.join(self.test_samples_path, "invalid_lex", "backtick.c")
@@ -185,7 +185,7 @@ class TestLexerChapter01(unittest.TestCase):
         with NamedTemporaryFile(suffix=".i") as preprocessed_file:
             gcc_preprocess(path, preprocessed_file.name)
             with self.assertRaises(ValueError):
-                lexer.run_lexer(preprocessed_file.name)
+                lexer.tokenize(preprocessed_file.name)
 
     def test_invalid_identifier_2(self):
         path = os.path.join(
@@ -195,7 +195,7 @@ class TestLexerChapter01(unittest.TestCase):
         with NamedTemporaryFile(suffix=".i") as preprocessed_file:
             gcc_preprocess(path, preprocessed_file.name)
             with self.assertRaises(ValueError):
-                lexer.run_lexer(preprocessed_file.name)
+                lexer.tokenize(preprocessed_file.name)
 
     def test_invalid_identifier(self):
         path = os.path.join(
@@ -205,7 +205,7 @@ class TestLexerChapter01(unittest.TestCase):
         with NamedTemporaryFile(suffix=".i") as preprocessed_file:
             gcc_preprocess(path, preprocessed_file.name)
             with self.assertRaises(ValueError):
-                lexer.run_lexer(preprocessed_file.name)
+                lexer.tokenize(preprocessed_file.name)
 
 
 if __name__ == "__main__":
